feat：LLM 失败重试逻辑公共化 + free_dialogue 读档续轮——①generate_with_retry 抽到 events/mod.rs 供 ai_dialogue/free_dialogue 复用（自动重试 3 次 + 重试按钮，不踢出剧本）②free_dialogue 把已完成轮次写入 vars（key 带事件索引），读档续跑从下一轮继续：不重置轮次、不重复已聊内容、不重放整块
