+++ src-tauri/src/ai_service/game_system/script_engine/events/ai_dialogue_event.rs
@@ -6,13 +6,9 @@ use serde_json::Value;
 use tauri::Manager;
 
 use crate::ai_service::game_system::script_engine::events::{
-    parse_duration, register_event, wait_for_frontend_continue, ScriptContext, ScriptEvent,
-};
-use crate::ai_service::game_system::script_engine::responses::{
-    event_names::SCRIPT_LLM_RETRY, LlmRetryPayload,
+    generate_with_retry, parse_duration, register_event, ScriptContext, ScriptEvent,
 };
 use crate::ai_service::game_system::script_engine::utils::script_function;
-use crate::ai_service::message_system::events::emit;
 use crate::ai_service::message_system::generator::{
     GeneratorDeps, GeneratorSource, MessageGenerator,
 };
@@ -116,47 +112,10 @@ impl ScriptEvent for AIDialogueEvent {
 
         let generator = MessageGenerator::new(deps);
 
-        // LLM 调用失败**绝不踢出玩家**（剧本连续性优先，不跳过对话、不退出剧本）：
-        // 1) 先自动重试 3 次（退避约 1s/2s/3s，覆盖瞬时网络抖动）；
-        // 2) 仍失败则广播「重试」提示并等待玩家点击「继续」——玩家点继续经
-        //    `script_event_continue` 回执到 continue_tx（与旁白共用等待通道，
-        //    引擎顺序执行无冲突），此后重置自动重试计数再来一轮；
-        // 3) 玩家不想等可退出剧本：stop_script abort 本任务即终止。
-        // 重试回溯点正确性：失败时 process_next_event 未返回（进度仍停在当前
-        // ai_dialogue）、add_assistant_line 未执行（line_list 无残留），重新生成
-        // 的上下文与首次一致。
-        let mut attempt: u64 = 0;
-        loop {
-            match generator.process_message(None).await {
-                Ok(_) => break,
-                Err(e) => {
-                    attempt += 1;
-                    tracing::warn!(
-                        "[AIDialogueEvent] LLM 生成失败（第 {} 次重试）: {}",
-                        attempt,
-                        e
-                    );
-                    if attempt > 3 {
-                        // 自动重试耗尽：交还玩家控制
-                        let _ = emit(
-                            ctx.app,
-                            SCRIPT_LLM_RETRY,
-                            &LlmRetryPayload {
-                                message: format!(
-                                    "AI 响应失败（已尝试 {} 次），点击「继续」重试",
-                                    attempt
-                                ),
-                            },
-                        );
-                        wait_for_frontend_continue(&ctx.channels).await;
-                        attempt = 0;
-                        continue;
-                    }
-                    // 自动重试退避（约 1s/2s/3s）
-                    tokio::time::sleep(std::time::Duration::from_secs(attempt as u64)).await;
-                }
-            }
-        }
+        // LLM 调用失败**绝不踢出玩家**：自动重试 3 次后广播「重试」提示、等玩家点
+        // 「继续」再试（详见 generate_with_retry 的注释：不跳过对话、不退出剧本，
+        // 重试回溯点正确——进度停在当前事件、line_list 无残留）。
+        generate_with_retry(ctx, &generator).await?;
 
         tracing::info!("[AIDialogueEvent] 执行完毕");
